+++ CHANGELOG.md
@@ -7,5 +7,5 @@ First release.
 
 ## 0.1.1.0 - 2023-09-18
 
-New data type `IntegralResult` to store the result (value, error, code) instead of a triplet.
+New data type `IntegralResult` to store the result (value, error, code) instead of in a triplet.