+++ CHANGELOG.md
@@ -7,5 +7,6 @@ First release.
 
 ## 0.1.1.0 - 2023-09-18
 
-New data type `IntegralResult` to store the result (value, error, code) instead of in a triplet.
+- New data type `IntegralResult` to store the result (value, error, code), instead of in a triplet.
 
+- Replaced `Double` with `CDouble` and `Int` with `CInt`.